source/moqawalatiWeb/src/test/java/dz/alkhwarizmix/framework/java/dtos/security/model/vo/PasswordTest.java
////////////////////////////////////////////////////////////////////////////////
//  بسم الله الرحمن الرحيم
//
//  حقوق التأليف والنشر ١٤٣٦ هجري، فارس بلحواس (Copyright 2015 Fares Belhaouas)  
//  كافة الحقوق محفوظة (All Rights Reserved)
//
//  NOTICE: Fares Belhaouas permits you to use, modify, and distribute this file
//  in accordance with the terms of the license agreement accompanying it.
//
////////////////////////////////////////////////////////////////////////////////

package dz.alkhwarizmix.framework.java.dtos.security.model.vo;

import static org.junit.Assert.assertEquals;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.runners.MockitoJUnitRunner;

/**
 * <p>
 * TODO: Javadoc
 * </p>
 * 
 * @author فارس بلحواس (Fares Belhaouas)
 * @since ١٥ ربيع الثاني ١٤٣٦ (February 04, 2015)
 */
@RunWith(MockitoJUnitRunner.class)
@SuppressWarnings("PMD.MethodNamingConventions")
public class PasswordTest {

	// --------------------------------------------------------------------------
	//
	// Setup & Teardown
	//
	// --------------------------------------------------------------------------

	@InjectMocks
	private Password utPassword;

	// --------------------------------------------------------------------------
	//
	// Tests
	//
	// --------------------------------------------------------------------------

	@Test
	public void test01_set_then_get_encryption() {
		Encryption value = new Encryption("Encryption1");
		utPassword.setEncryption(value);
		assertEquals(value, utPassword.getEncryption());
	}

	@Test
	public void test02_set_then_get_password() {
		String value = "Password1";
		utPassword.setPassword(value);
		assertEquals(value, utPassword.getPassword());
	}

	@Test
	public void test03_set_then_get_user() {
		User value = new User("User1");
		utPassword.setUser(value);
		assertEquals(value, utPassword.getUser());
	}

} // Class
